Removal of bot_action.py
